gcloud_functions/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-vision
google-cloud-storage
google-cloud-texttospeech
